# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Salon {
  id: number;
  name: string;
  area: string;
  address: string;
  rating: number;
  reviewCount: number;
  imageUrl: string;
  categories: string[];
  priceRange: string;
  openNow: boolean;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  openingTime?: string | null;
  /** @nullable */
  closingTime?: string | null;
  featured?: boolean;
}

export interface Service {
  id: number;
  salonId: number;
  name: string;
  category: string;
  price: number;
  duration: number;
  /** @nullable */
  description?: string | null;
  popular?: boolean;
}

export interface Review {
  id: number;
  salonId: number;
  customerName: string;
  rating: number;
  comment: string;
  createdAt: string;
  /** @nullable */
  serviceName?: string | null;
}

export type SalonDetail = Salon & {
  services?: Service[];
  reviews?: Review[];
  gallery?: string[];
};

export interface SalonInput {
  name: string;
  area: string;
  address: string;
  phone: string;
  description: string;
  priceRange: string;
  openingTime: string;
  closingTime: string;
  imageUrl: string;
  categories?: string[];
}

export interface ServiceInput {
  salonId: number;
  name: string;
  category: string;
  price: number;
  duration: number;
  description?: string;
}

export interface Booking {
  id: number;
  salonId: number;
  salonName: string;
  serviceId: number;
  serviceName: string;
  customerName: string;
  customerPhone: string;
  appointmentDate: string;
  appointmentTime: string;
  status: string;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface BookingInput {
  salonId: number;
  serviceId: number;
  customerName: string;
  customerPhone: string;
  appointmentDate: string;
  appointmentTime: string;
  notes?: string;
}

export interface BookingStatusUpdate {
  status: string;
}

export interface ReviewInput {
  customerName: string;
  rating: number;
  comment: string;
  serviceName?: string;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  salonCount: number;
}

export interface Area {
  name: string;
  salonCount: number;
}

export interface MarketplaceStats {
  totalSalons: number;
  totalServices: number;
  totalBookings: number;
  avgRating: number;
  topAreas?: Area[];
}

export type ListSalonsParams = {
category?: string;
area?: string;
search?: string;
minRating?: number;
};

export type ListServicesParams = {
category?: string;
};

export type ListBookingsParams = {
salonId?: number;
customerName?: string;
};