# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  salonId: number;
  salonName: string;
  serviceId: number;
  serviceName: string;
  customerName: string;
  customerPhone: string;
  appointmentDate: string;
  appointmentTime: string;
  status: string;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}