# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  salonId: number;
  serviceId: number;
  customerName: string;
  customerPhone: string;
  appointmentDate: string;
  appointmentTime: string;
  notes?: string;
}