# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Area } from './area';

export interface MarketplaceStats {
  totalSalons: number;
  totalServices: number;
  totalBookings: number;
  avgRating: number;
  topAreas?: Area[];
}