# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  salonId: number;
  customerName: string;
  rating: number;
  comment: string;
  createdAt: string;
  /** @nullable */
  serviceName?: string | null;
}