# lib/api-zod/src/generated/types/salon.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Salon {
  id: number;
  name: string;
  area: string;
  address: string;
  rating: number;
  reviewCount: number;
  imageUrl: string;
  categories: string[];
  priceRange: string;
  openNow: boolean;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  openingTime?: string | null;
  /** @nullable */
  closingTime?: string | null;
  featured?: boolean;
}