# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Beauty Salon Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  salonId: number;
  name: string;
  category: string;
  price: number;
  duration: number;
  /** @nullable */
  description?: string | null;
  popular?: boolean;
}